lldpad: Add '-p' option to surpress PID file creation
A PID file is not always necessary, so add option '-p' to
suppress PID file handling.

+++ lldpad.c
@@ -122,7 +122,7 @@ static void usage(void)
 {
 	fprintf(stderr,
 		"\n"
-		"usage: lldpad [-hdksv] [-f configfile]"
+		"usage: lldpad [-hdkspv] [-f configfile]"
 		"\n"
 		"options:\n"
 		"   -h  show this usage\n"
@@ -130,6 +130,7 @@ static void usage(void)
 		"   -d  run daemon in the background\n"
 		"   -k  terminate current running lldpad\n"
 		"   -s  remove lldpad state records\n"
+		"   -p  Do not create PID file\n"
 		"   -v  show version\n"
 		"   -V  set syslog level\n");
 
@@ -220,16 +221,17 @@ int main(int argc, char *argv[])
 {
 	int c;
 	struct clif_data *clifd;
-	int fd;
+	int fd = -1;
 	char buf[32];
 	int shm_remove = 0;
 	int killme = 0;
 	int print_v = 0;
+	int pid_file = 1;
 	pid_t pid;
 	int cnt;
 
 	for (;;) {
-		c = getopt(argc, argv, "dhkvsf:V:");
+		c = getopt(argc, argv, "dhkvspf:V:");
 		if (c < 0)
 			break;
 		switch (c) {
@@ -248,6 +250,9 @@ int main(int argc, char *argv[])
 			break;
 		case 's':
 			shm_remove = 1;
+			break;
+		case 'p':
+			pid_file = 0;
 			break;
 		case 'v':
 			print_v = 1;
@@ -323,25 +328,27 @@ int main(int argc, char *argv[])
 		exit (0);
 	}
 
-	fd = open(PID_FILE, O_RDWR | O_CREAT, S_IRUSR | S_IWUSR);
-	if (fd < 0) {
-		LLDPAD_ERR("error opening lldpad lock file");
-		exit(1);
-	}
-
-	if (flock(fd, LOCK_EX | LOCK_NB) < 0) {
-		if (errno == EWOULDBLOCK) {
-			fprintf(stderr, "lldpad is already running\n");
-			if (read(fd, buf, sizeof(buf)) > 0) {
-				fprintf(stderr, "pid of existing lldpad is %s\n",
-					buf);
-			}
-			LLDPAD_ERR("lldpad already running");
-		} else {
-			perror("error locking lldpad lock file");
-			LLDPAD_ERR("error locking lldpad lock file");
+	if (pid_file) {
+		fd = open(PID_FILE, O_RDWR | O_CREAT, S_IRUSR | S_IWUSR);
+		if (fd < 0) {
+			LLDPAD_ERR("error opening lldpad lock file");
+			exit(1);
 		}
-		exit(1);
+
+		if (flock(fd, LOCK_EX | LOCK_NB) < 0) {
+			if (errno == EWOULDBLOCK) {
+				fprintf(stderr, "lldpad is already running\n");
+				if (read(fd, buf, sizeof(buf)) > 0) {
+					fprintf(stderr, "pid of existing"
+						"lldpad is %s\n", buf);
+				}
+				LLDPAD_ERR("lldpad already running");
+			} else {
+				perror("error locking lldpad lock file");
+				LLDPAD_ERR("error locking lldpad lock file");
+			}
+			exit(1);
+		}
 	}
 
 	lldpad_oom_adjust();
@@ -375,36 +382,41 @@ int main(int argc, char *argv[])
 		goto out;
 	}
 
-	if (lseek(fd, 0, SEEK_SET) < 0) {
-		LLDPAD_ERR("error seeking lldpad lock file\n");
-		exit(1);
+	if (pid_file) {
+		if (lseek(fd, 0, SEEK_SET) < 0) {
+			LLDPAD_ERR("error seeking lldpad lock file\n");
+			exit(1);
+		}
+
+		memset(buf, 0, sizeof(buf));
+		sprintf(buf, "%u\n", getpid());
+		if (write(fd, buf, sizeof(buf)) < 0)
+			perror("error writing to lldpad lock file");
+		if (fsync(fd) < 0)
+			perror("error syncing lldpad lock file");
+
+		close(fd);
 	}
-
-	memset(buf, 0, sizeof(buf));
-	sprintf(buf, "%u\n", getpid());
-	if (write(fd, buf, sizeof(buf)) < 0)
-		perror("error writing to lldpad lock file");
-	if (fsync(fd) < 0)
-		perror("error syncing lldpad lock file");
-
-	close(fd);
 
 	pid = lldpad_shm_getpid();
 	if (pid < 0) {
 		LLDPAD_ERR("error getting shm pid");
-		unlink(PID_FILE);
+		if (pid_file)
+			unlink(PID_FILE);
 		exit(1);
 	} else if (pid == PID_NOT_SET) {
 		if (!lldpad_shm_setpid(getpid())) {
 			perror("lldpad_shm_setpid failed");
 			LLDPAD_ERR("lldpad_shm_setpid failed\n");
-			unlink(PID_FILE);
+			if (pid_file)
+				unlink(PID_FILE);
 			exit (1);
 		}
 	} else if (pid != DONT_KILL_PID) {
 		if (!kill(pid, 0)) {
 			LLDPAD_ERR("lldpad already running");
-			unlink(PID_FILE);
+			if (pid_file)
+				unlink(PID_FILE);
 			exit(1);
 		}
 		/* pid in shm no longer has a process, go ahead
@@ -413,7 +425,8 @@ int main(int argc, char *argv[])
 		if (!lldpad_shm_setpid(getpid())) {
 			perror("lldpad_shm_setpid failed");
 			LLDPAD_ERR("error overwriting shm pid");
-			unlink(PID_FILE);
+			if (pid_file)
+				unlink(PID_FILE);
 			exit (1);
 		}
 	}
@@ -456,8 +469,9 @@ int main(int argc, char *argv[])
 	if (ctrl_iface_register(clifd) < 0) {
 		if (!daemonize)
 			fprintf(stderr, "failed to register control interface\n");
-		LLDPAD_ERR("lldpad failed to start - failed to register control interface\n");
-		exit(1);
+		LLDPAD_ERR("lldpad failed to start - "
+			   "failed to register control interface\n");
+		goto out;
 	}
 
 	eloop_run();
@@ -471,7 +485,8 @@ int main(int argc, char *argv[])
 out:
 	destroy_cfg();
 	closelog();
-	unlink(PID_FILE);
+	if (pid_file)
+		unlink(PID_FILE);
 	eloop_destroy();
 	if (eloop_terminated())
 		exit(0);
